LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-es-419-expected.txt

PASS navigator.language is "es-419"
PASS successfullyParsed is true

TEST COMPLETE
